// client-app/src/app/models/service/serviceSpecification.ts
export interface ServiceSpecification {
    makeId: string;
    makeDescription?: string;
    modelDescription?: string;
    fromDate: Date;
    serviceSpecificationId?: string;
    productId: any;
    modelId: string;
    standardTimeInMinutes: number;
    productName?: string;
    thruDate?: any;
}